refactor: split commandViewerComponent
